[TASK] Move files (Related to #2)
