+++ Configuration/TypoScript/constants.txt
@@ -1,0 +1,6 @@
+
+#cat=plugin_tt_news_opengraph/image/010; type=image; label=Image: Default image if no image is available;
+plugin.tt_news.tx_jhopengraphttnews.nopic_path = EXT:jh_opengraph_ttnews/Resources/Public/Images/nopic.jpg
+
+#cat=plugin_tt_news_opengraph//020; type=string; label=Enable for: Enable open graph protocol for view (comma-separated list);
+plugin.tt_news.tx_jhopengraphttnews.enableForWhatToDisplay = SINGLE
