+++ Configuration/TypoScript/setup.txt
@@ -1,0 +1,6 @@
+config.htmlTag_setParams =	xmlns:og="http://ogp.me/ns#" xmlns:fb="https://www.facebook.com/2008/fbml"
+
+plugin.tt_news.tx_jhopengraphttnews {
+    nopic_path = {$plugin.tt_news.tx_jhopengraphttnews.nopic_path}
+    enableForWhatToDisplay = {$plugin.tt_news.tx_jhopengraphttnews.enableForWhatToDisplay}
+}
