+++ ext_tables.php
@@ -1,0 +1,5 @@
+<?php
+if (!defined ('TYPO3_MODE'))
+    die ('Access denied.');
+
+TYPO3\CMS\Core\Utility\ExtensionManagementUtility::addStaticFile($_EXTKEY,'Configuration/TypoScript/','News open graph protocol');
